Introduced Role Based access

+++ src/main/java/com/haufe/beercatalogue/service/BeerService.java
@@ -17,10 +17,16 @@ import com.haufe.beercatalogue.repository.ManufacturerRepository;
 public class BeerService {
     private final BeerRepository beerRepository;
     private final ManufacturerRepository manufacturerRepository;
+    private final AccessService accessService;
 
-    public BeerService(final BeerRepository beerRepository, final ManufacturerRepository manufacturerRepository) {
+    public BeerService(
+            final BeerRepository beerRepository,
+            final ManufacturerRepository manufacturerRepository,
+            final AccessService accessService
+    ) {
         this.beerRepository = beerRepository;
         this.manufacturerRepository = manufacturerRepository;
+        this.accessService = accessService;
     }
 
     @Transactional(readOnly = true)
@@ -41,21 +47,25 @@ public class BeerService {
 
     public Beer create(final Beer beer) {
         beer.setManufacturer(getManufacturer(beer));
+        accessService.requireAdminOrOwnManufacturer(beer.getManufacturer().getId());
         return beerRepository.save(beer);
     }
 
     public Beer update(final Long id, final Beer beer) {
         final var existingBeer = findById(id);
+        accessService.requireAdminOrOwnManufacturer(existingBeer.getManufacturer().getId());
         existingBeer.setName(beer.getName());
         existingBeer.setAbv(beer.getAbv());
         existingBeer.setType(beer.getType());
         existingBeer.setDescription(beer.getDescription());
         existingBeer.setManufacturer(getManufacturer(beer));
+        accessService.requireAdminOrOwnManufacturer(existingBeer.getManufacturer().getId());
         return beerRepository.save(existingBeer);
     }
 
     public void delete(final Long id) {
         final var beer = findById(id);
+        accessService.requireAdminOrOwnManufacturer(beer.getManufacturer().getId());
         beerRepository.delete(beer);
     }
